Put utils functions in the same file
