+++ src/pixelator/pna/cli/graph.py
@@ -23,7 +23,7 @@ from pixelator.pna.graph.community_detection import (
     StagedRefinementOptions,
 )
 from pixelator.pna.graph.component_recovery import build_pxl_file_with_components
-from pixelator.pna.graph.connected_components_common import ConnectedComponentException
+from pixelator.pna.graph.component_recovery_utils import ConnectedComponentException
 from pixelator.pna.graph.constants import MIN_PNA_COMPONENT_SIZE
 from pixelator.pna.graph.report import GraphSampleReport
